BUG#35595 : Backup crash server
The crash described in the bug report cannot be repeated.
This patch contains a regression test for the command that
caused the original crash. Note that there are two tests
because the output file name of '*' has different outcomes
depending on the platform chosen.
      
These tests can also be the basis for further tests for
BUG#40475.