Bug#43090 - ARCHIVE engine related warnings on Windows debug build
A windows debug build created warnings in the error log.
In a windows debug build, we do not define SAFEMALLOC and SAFE_MUTEX.
But the ARCHIVE engine explicitly defined them in CMakeLists.txt.
Ha_archive.cc uses my_multi_malloc(). This is part of mysys, which,
like most parts of the server, are compiled without SAFEMALLOC.
So my_multi_malloc() calls my_malloc() directly and is not
redirected to the SAFEMALLOC function _mymalloc(). But since
ha_archive.cc frees the allocated memory with my_free(), which
is redefined to _myfree(), SAFEMALLOC complains about a free, for
which it didn't see an allocation.

Fixed by not explicitly defining SAFEMALLOC and SAFE_MUTEX in
ARCHIVE's CMakeLists.txt.