Bug#42340 - RelordLeaf::pruneRecords() was hitting the assertion (oldestVisible->state != recLock).   This was always a lock record at savepoint 0 when the more recent record was recDeleted.
It is surprisingly OK for a committed record to have left over lock records
in the prior record chain.  This happens when an updated or deleted record
version is at a savepoint higher than the lock record. If a commit happens 
instead of an explicit releae savepoint,the call to Transaction::releaseSavepoint() 
is not made, which would have reduced the number of records on that same 
transaction to 1.

Instead, Transaction::commit() calls Transaction::releaseSavepoints().  This 
function just clears out the savepoint objects and does not try to scavenge any 
records from the prior record chain.  So they are left on there for the scavenger
to clean up later.

The new scavenger did not know this and assumed that a lock record left in the
chain after a commit is wrong.  It is OK, so the ASSERT is herby deleted.

In addition, RecordScavenger::inventoryRecord() is modfied to be smarter about
where it indicates that pruning can start.  Since the record version newer than the 
lock record is a real record, it is usually the one designated as the oldest visible.
But if the timing is right, the lockRecord may get a temporary useCount by another
thread.  If inventoryRecord looks at it just then, it would adjust the 'oldestVisible'
designation to this lock record.  This is now changed to set oldestVisible to NULL.
Then the next prior record will get that designation if it exists.

Also, at the point where the oldestVisible is assigned, the ASSERT that it is not a
lockRecord is changed to an 'if' since we now know that lockRecords can occur in the 
prior record chain.