Bug#41742 - Since Table::fetchNext seems to do OK if a bit below the higWater mark is set incorrectly, but does not do OK if a bit below highWater is cleared incorrectly, I am changing Table::insertIntoTree so that it sets bits but does not clear them. Bits are already not cleared by RecordLeaf::retireRecords.
So it will be the responsibility of Table::fetchRecords to clear bits below the highWater 
mark if the record is not found in either the record tree or the database.  This is done 
under an exclusive lock of Table::syncObject. After the lock is made and before we clear 
that bit, another fetch is attempted just in case another thread just added that record.

Since Table::insertIntoTree no longer clears bits in recordBitmap, we no longer need to 
re-set the bit in Table::backlogRecord().