WL#4630 - ST: MySQL Backup Client Program - Milestone 1
The MySQL Backup Client Program is a command line client
program, that examines and displays the contents of a
backup image.

By command line options you can select different types
of information to view.

It does also have a search capability for finding
specific database objects in the backup image.

For a detailed explanation of the options, run

mysqlbackup --help