Bug#41578 - Drop column/table with grants followed by restore fails.
When a database object like a table or column is dropped,
privileges for these objects are not dropped.

BACKUP includes all privileges that belong to the saved databases.
So it includes privileges for objects that do not exist.

On RESTORE the saved objects are re-created, followed by the
privileges. RESTORE failed when trying to grant a privilege
for a non-existent object.

To be able to restore the same objects and privileges as they existed
at backup time, we do now omit checks for object existence when
granting privileges during RESTORE.