Fix for BUG#40633 Falcon assertion "lockState == 0" failed in SyncObject.cpp
The cause for this crash was that when running the embedded test
suite, Falcon is just started and then for many test stopped almost
immediately. This could result in that the IO threads never started to
run. As a result the Cache could be shutdown with a pending flush
request and the corresponding syncWait lock locked.

This fix solves this problem by waiting for the IO threads to start
before returning from the initialization of the Page Cache.