Code cleanup: remove occurences of priorPage in the code.
IndexPage::piorPage was not really used. It was maintained for some future (or past) functionality, according to rumors to consolidate pages. Still the code to maintain prior page pointers was wrong with respect to recovery.

This patch removes all occurences of IndexPage::priorPage, except in structure definition, to keep persistent on disk structures compatible. Maybe these 4 bytes  can be used in the future for some purpose.