Bug#42344  Falcon recovery Exception: read error on page Bug#42903  Falcon: Inconsistent tablespace after recovery
The exception is caused during recovery of an index page, when 
trying to  insert a first record to the parent page of the current 
page. Parent page was not existing - this caused reads behind the 
EOF of falcon_user.fts 

The problem is that parent page itself was not logged.
Another (worse) problem that can result into corrupted index is that 
that parent pointer is not reliable and often can be wrong for 
non-leaf pages. When non-leaf page is split, half or its nodes are 
transferred to the newly created page. For  consistency, parentPage
of corresponding child pages had to be changed and change had to be 
logged and this was not done.

The fix is to remove all occurenced parentPage pointer from the 
Falcon code. Having a link to parent page was an idea that is not 
implemented consistently. Also, this patch removed all occurenced 
to priorPage from the Falcon code.This was another inconsistently 
implemented idea.

IndexPage::priorPage and IndexPage::parentPage are renamed to  
int32 IndexPage::unused[2], for future use and to prevent the need
to upgrade for existing databases.

Recovery is changed to perform careful writes in the right order 
during the index split (newly created page, then page that overflowed, 
then parent page or the newly creted page).

Doing recovery like this could result in a slightly detuned index, if 
crash or kill occures before parent page is logged. Detuned means
there is a page without a parent. Index traversal will still find it
when traversing the page level via IndexPage::next links. 

"Detuned" problem can be fixed  by logging  all pages involved in index 
split at once. This is something to consider in the future.