Bug#41979 - Routine level grants not restored when user is dropped,             recreated before restore
BACKUP/RESTORE did not implement backup and restore of privileges
for stored procedures and stored functions.

On BACKUP, we do now select routine privileges from mysql.proc_priv.
This is a preliminary solution until Bug 26886 / WL 2646 are implementd.

Another BACKUP change is in the GRANT statement creation. Grants for
routines have the additional keyword PROCEDURE or FUNCTION after ON.

On RESTORE, we need to be aware of the additional keyword, when we
search for the database name in the GRANT statement.