Fix to Bug #40155 Error: assertion (ret == 0) failed at line 80 in file Mutex.cpp
This crash was due to the Memory manager trying to lock the same mutex twice. It happened
when the server run out of memory. The memory manager handles this by throwing an exception.
Unfortunatly, when creating this exception there there was call both the JString and the
Log class that did allocate memory leading to the second call to the memory manager's
mutex.
      
This patch fixes this by:
      
-SQLError use an new method on JString that does not allocate memory for the text
 but just stores the pointer to the existing string
-Simplified SQLError's use of the Log if the error is out-of-memory