BUG#41642 : The purpose for the backup_engines suite isn't clear.
This patch reorganizes the backup_engines tests. A new suite is created
for those tests that must iterate over replication modes (backup_ptr). 
The storage engines list is extended to include all storage engines 
except for MERGE and FEDERATED which require special commands to utilize. 

A description for each of the suites has been added to clarify the usage 
of the suites and what sort of tests should be included.

Note: The archive storage engine is disabled for the partition tests. 
See BUG#43028 and BUG#43030 for details.

The following are commands needed to apply this patch. These create the 
new folders and move the test files. The changes to the test files are 
included in the patch. These should be executed from the root of the 
source tree.

Added backup_ptr to the default list in new mtr.