This is a patch for bug#43452 Wrong number of rows returned when using LIMIT
In a transient period, we have both the old and the new values available when walking the
index. The order they are traversed is the following:    (notation  (recNumber,
indexValue)  )

(1,1) (0,3) (1,3) (0,8)

We check the first value in IndexWalker::getValidatedRecord and see that the index value
(1) does not match the record value (3), therefore we skip it. At the next iteration we
check the next entry. We check that this is not the same record as the on the last
iteration by comparing the record number to a variable called lastRecordNumber. This value
is initialized to 0, and has not yet been updated because we skipped the previous entry.
Therefore we mistakingly think we have processed this record already and skip it. We
therefore loose the first record in the table.

The solution to this is not to initialize lastRecordNumber to 0, which is a legal record
number, but to initialize it to a negative, illegal record number.

Added file 'mysql-test/suite/falcon/t/falcon_bug_43452.test'
------------------------------------------------------------
A test testing the patch

Added file 'mysql-test/suite/falcon/r/falcon_bug_43452.result'
-------------------------------------------------------------
Expected result of the test

Modified file 'storage/falcon/IndexWalker.cpp'
----------------------------------------------
Initialize lastRecordNumber to a negative value