Bug#43500 - Backup: too many tests require debug compile
Many backup test files require debug builds to execute even though the tests themselves do not rely on debug functionality. This is bad because these tests will only be run on servers with debug builds in pushbuild.
      
With this patch, the requirement for a debug built server is removed from backup test files that do not rely on debug functionality.

Also deleted backup_view_on_view. This file seems to live it's own life. I have deleted it at least two times before (the test is incorporated into backup_views), but it persistently returns from the land of the dead every time.