This is a patch for Bug #43488 Crash in Table::fetchForUpdate (update conflict) when using LIMIT query
This crash was due to missing exception handling when calling fetchForUpdate
when fetching records through the IndexWalker. The exception handling
introduced in this patch is split in two parts. The first part does the 
necessary cleanup in IndexWalker::getValidatedRecord and re-throws the 
exception. The exception is re-caught in StorageDatabase::nextIndexed, 
where we return the correct error.

I have not included any test files for this patch, because it is not
easily reproducible without a larger framework as described in the bug
report.

Modified file 'storage/falcon/IndexWalker.cpp'
----------------------------------------------
Added try - catch logic, doing the necessary cleanup in the case
of an exception. This implementation is based on the similar cases
in StorageDatabase.cpp. After the cleanup has been done, we re-throw
the exception, so that we can return the appropriate error in 
StorageDatabase::nextIndexed
 
Modified file 'storage/falcon/StorageDatabase.cpp'
--------------------------------------------------
Recatch the exception and return the appropriate error.