Bug#39063 - Online Backup: Backup behavior changes with the case used for the database name
Before, the database list for BACKUP was case sensitive. This caused problems in
case insensitive servers because database 'X' and 'x' is considered the same database
while backup would fail if the wrong case was used.
                  
With this patch, the database names are converted to lower case when BACKUP is
executed on a case insensitive server.