BUG#43075 rpl.rpl_sync fails sporadically on pushbuild
The slave was crashing while failing to execute the init_slave() function.

The issue may stem from two different reasons:

  1 - A failure while allocating the master info structure generated a
      segfault due to a NULL pointer.

  2 - A failure while recovering generated a segfault due to a non-initialized
      relay log file. In other words, the mi->init and rli->init were both set to true
      before executing the recovery process thus creating an inconsistent state as the
      relay log file was not initialized.

To circumvent such problems, we refactored the recovery process which is now executed while
initializing the relay log. It is ensured that the master info structure is created
before accessing it and any error is propagated thus avoiding to set mi->init and
rli->init to true when for instance the relay log is not initialized or the relay info
is not flushed.

The changes related to the refactory are described below:

  1 - Removed call to init_recovery from init_slave.

  2 - Changed the signature of the function init_recovery.

  3 - Removed flushes. They are called while initializing the relay log and master info.

  4 - Made sure that if the relay info is not flushed the mi-init and rli-init are not
  set to true.

In this patch, we also replaced the exit(1) in the fault injection by DBUG_ABORT() to
make it compliant with the code guidelines.