BUG#35079/WL#4538 - Make BACKUP/RESTORE statements interruptible.
The server detects interruptions in statement execution such as when client 
connection is closed or if user hits Ctrl+C. However, the code inside the server 
must actively check if an interruption has happened and abort execution in that 
case.
      
This patch adds such checks to the backup code, as described in WL#4538. After 
this patch it should be possible to interrupt on-going BACKUP/RESTORE command. 
It also fixes few problems in the code so that the new tests can pass through.