Bug#41722 - Restore crashes with backup of Falcon tables
During RESTORE of a not so small Falcon table, the server crashed.

The cause for the crash was an attempt to retry to send table data
to the restore driver after an error from the restore driver.
Restore::send_data() does not expect to be called again with
table data after returning an error.

Fixed by aborting RESTORE, if a driver returns an error.