A patch for Bug#35297: SHOW CREATE EVENT does not show the DEFINER.
The problem was that SHOW CREATE EVENT output does not include
the DEFINER clause, even if is specified in the CREATE EVENT statement.
SHOW CREATE EVENT is used for backup/restore metadata, so it is essential
to have the DEFINER clause in the output.

The fix is to include DEFINER clause in the generated CREATE EVENT statement.

mysqldump needs to enclose the clause in version-specific comment in order
to generate the dump that is compatible with old server versions
(the versions which don't support DEFINER in events).