A patch for Bug#42384 (thread_handling = pool-of-threads, wrong handling of max_questions user limit).
The problem was that in case of pool-of-threads a thd object
was not properly initialized, i.e. thr_create_time and start_utime
were not set. So, time_out_user_resource_limits(), that is
essential for handling user limits, was not able to work properly.

The fix is to initialize forgotten attributes properly.