Fix for Bug #41665 Falcon crash in Transaction::waitForTransaction
This crash occurred in the deadlock detector code. The current thread
A is traversing the waitingFor list while another thread C has just
aborted (or possibly committed) and is releasing its reference count
on the transaction state object. A third thread, B, that is just in
front of C in the waitingFor list is waken up, and also releases its
reference count on C. If A is inspecting object C as this happens,
this object is deleted - and can result in thread A crashing.

The fix to Transaction::waitForTransaction(): Before releasing the
reference count on the transaction state object we aquire an exclusive
lock on the active transaction list.  This will avoid that the
transaction state object is released and deleted while another thread
is transversing the waitingFor list.

Also added a missing call to transState->release() in the catch block.