Bug #42131, "falcon_backlog test crashes in Record::getEncodedValue" Bug #38375, "Falcon crash in Record::isNull"
For sequential operations such as index and table scans, StorageTable
caches the current record in StorageTable::record. In rare cases, it is
possible that StorageTable::record may have been chilled. This is a
problem if an attempt is made to examine the contents of the record.

Calls from the server directly into the storage interface, such as
StorageInterface::rnd_next, bypass the internal checks in Falcon that
ensure a record is thawed before examining it.To avoid crashes in this
circumstance, StorageTable::compareKey() now thaws StorageTable::record
if necessary before examining it.

Note that it is not necessary to always thaw StorageTable::record, for
example in StorageTable::setRecord(), because the contents of the
record is not always examined as it is in compareKey().

If a thaw is necessary, and if the thaw fails, then compareKey() throws
an INTERNAL_ERROR exception.