Bug#43363 - Backup: Backup on case sensitive server restore on case insensitive server fails
Before, RESTORE on a case insensitive server would fail if the backup image contained databases or tables with upper case names. 
      
Now, RESTORE is able to handle this case by converting the names to lowercase in the restore catalog as long as there are no duplicate names after the conversion. The duplicate name problem is logged as bug 43596.