Bug#43958 - An alternative solution to deleting the assert for the existence of the TransState object is to re-organize the calls to Transaction::waitForTransaction so that they ALWAYS use a TransactionState pointer.  This can be done because the only place where the transId is sent in by itself is in Table::validateAndInsert().  This function calls waitForTransaction when a new base record is found superceding the record that was known as the prior.  This new base record MUST be a RecordVersion, so it MUST have a TransactionState pointer.  So instead of sending the TransId, the patch now sends the TransactionState pointer.  This also allows us to eliminate a costly and obnoxious search for a Transaction based on a TransId.
Note that Table::validateAndInsert() holds its own useCount on the TransactionState that it gets from the new pending base record, just in case that record is rolledBack.  Although, with the new CycleManager, that rolledBack record will not be deleted until validateAndInsert() is done.

In addition to waitForTransaction, the various calls to getRelativeState are also changed to use TransactionState and not use TransId. waitForTransaction does not need to increment its own useCount on the TransactionState object sent in.  It either comes from validateAndInsert or getRelativeState. Table::validateAndInsert holds a useCount itself.  Most of the getRelativeState calls hold a useCount on the record associated with the TransactionState, which by extention, holds the useCount on the TransactionState.  An addRef is added to getrelativeState(Record*) so that waitForTransaction does not need to.