Bug#43562 Crash when creating a Falcon table after upgrade from 6.0.9 to 6.0.10
The reason for the crash is that generation of multisegment key format was changed 
slightly as side-effect of fix for 42136. The new format worked ok with new databases
but not with the old ones. 

In this particular case, the error comes from Falcon's internal security mechanism. 
Due to changed index format, a query in Falcon's data dictionary unexpectedly returned 
an empty result set. This was interpreted was missing privilege to create tables.
Exception (security error) was thrown and not caught.

Fix:

- Define new minor ODS version 2.4 for databases with changed multisegment index.
- During the start, perform a check for 2.4 format(MySQL 6.10 would correspond to 
2.4 Falcon)

When building multisegment indexes, regard the ODS version and create different 
indexes in 2.3  and 2.4.