Bug#41971: Thread state on embedded server is always "Writing to net"
The problem is that the state of a thread on a embedded server is
always displayed as "Writing to net", which is wrong as there is
no "network" in the embedded server.

The solution is only exclude, on a embedded server, the thread
state conditions that are related to network operations. Other
thread states related to waiting on conditions or other operations
are preserved.