Bug#41661 - Table::insertIntoTree could fail during a Table::insert.  This should not happen since Dbb::insertStub should have returned a record number that was free.  Nobody else should have been messing with that slot in the RecordLeaf.
The problem was in RecordLeaf::retireRecords.  This function called RecordVersion::retire which 
then called expungeRecord. This makes the record number free to use again by any other thread.
It did this with an exclusive lock on the RecordLeaf, but Table::insertIntoTree often does not get 
any lock on RecordLeaf when it calls RecordLeaf::store.  So a free record number is reused in 
Table::insert before the slot has been set to NULL by RecordLeaf::retireRecords.  That is why 
Table::insertIntoTree fails.  It expects the slot to be NULL.

Another problem fixed is that the first of the two Table::insert functions did not check the 
return value of insertIntoTree at all.  This would have propogated the error further since 
there would be nothing in the RecordTree slot after the insert.

Another problem fixed in RecordLeaf::retireRecords is with this statement;
    if (record && recordScavenge->canBeRetired(record))
The problem is with incrementing the 'count' integer.  If it is zero after stepping through the 
RecordLeaf, then it is assumed that the RecordLeaf is empty and can be retired.
But this 'if' statement gets the wrong result.  If the RecordLeaf is full of newer records that
cannot yet be retired, then this count will remain zero.  The emptySections bitmap will be 
set for this RecordLeaf. And later, in Table::retireRecords, it will attempt to retire these 
sections.  But finding records in them, it will just waste a lot of time.  

I changed this code to be accurate in determining if the RecordLeaf is indead empty 
before setting the bit in Table::emptySections.

Record::retire and RecordVersion::retire now only retire records.  The more thorough 
tests for whether a record can be retired is collected into RecordScavenge::canBeRetired.

Also, instead of just setting the slot in RecordLeaf to NULL directly, I now use a 
COMPARE_EXCHANGE_POINTER() just like every other place where this location is changed.
Then if it fails for any reason, we will not retire that record.  One legitimate reason that this 
could fail is when this record is determined that it canBeRetired, but immediately after that, 
a client thread does a fetchForUpdate, adding a lockRecord.


Since the 'count' integer is incremented only within it, a RecordLeaf that does not have any