Bug #38173 Field doesn't have a default value with row-based replication
The reason of  the bug was incompatibile with the master side behaviour.
INSERT query on the master is allowed to insert into a table without specifying
values of DEFAULT-less fields if sql_mode is not strict.
      
Fixed with checking sql_mode by the sql thread to decide how to react.
Non-strict sql_mode should allow Write_rows event to complete.
      
todo: warnings can be shown via show slave status, still this is a 
      separate rather general issue how to show warnings for the slave threads.