Bug #43765 falcon_bug_22173a crashes in IndexPage::findNodeInLeaf
Falcon crashes with assertion (level==0)in high-concurrent
index updates

Reason is that recent code refactoring in IndexPage code introduced a 
regression.

When index is small and consists of a single root page only, findInsertionLeaf(), would check lock the root page with shared lock, check if the root page level is 0 (page is a leaf),  unlock the page, lock it again excusively and return this page. The problem is that during the gap while page was not locked, it could have beens spli. In this case returning root page is wrong, as it is not a "insertion leaf", and generally  not leaf page at all. 

The fix is not to return from the function prematurely, when reacquiring lock in the root and root page changes the level from 0 to greater number.