Bug#44155 - backup_intr_errors.test started to fail             after merge with main
backup_intr_errors.test failed to show warnings.

The cause was a forgotten (or lost by merge) switch
back to the default connection. Error injection was
requested on the wrong connection. After interrupt,
the respective warnings did not show up.

Unrelated change: Proper reset of the debug variable,
so that debugging by --debug is disturbed as little
as possible.