Bug#44068 - RESTORE can disable the MyISAM Key Cache
The test case myisam_keycache_coverage used to fail if it ran
after backup_myisam_sync.

The reason was that the latter test case disabled the key cache
during RESTORE. The error injection in myisam_keycache_coverage
became void as the key cache wasn't used any more. Reads of the
index file bypassed the cache, and the statements succeeded.

Fixed by moving initialization and de-initialization of the key
cache from mi_examine_log() to myisamlog. In the server, the
initialization is already done.

Additional change: To make the state of the key cache visible,
I added cleanup for the status variables blocks_used and
blocks_unused. This improves the testability.