BUG#43324 : Wrong table count per snapshot stored in backup image
The wrong table count is being written to the stream.
When the backup client reads the stream it produces the
wrong output. The table count always returned 1.

This is because the code was counting the tables in the
add_snapshot() method which is only called once.

This patch moves the table counting code to allow proper
counting of the tables in the snapshot.

As a result, the backup client test results have been
corrected as well.