Bug#38568 - This patch may help to fix this bug.  Needs testing.
All changes to Record::priorVersion are now done using 
COMPARE_EXCHANGE_POINTER instead of direct assignments so that 
we can be assured that a pointer was not lost.  If this fails, there are several 
new FATAL errors, which is better than continuing with a race condition.  
The only direct assignments left are in the RecordVersion constructors.

When pruning, a call is made to clearPriorVersion and the following code 
may or may not actually prune those records.  If not, they need to be 
re-attached with setPriorVersion.  That was a possible but very rare 
memory leak.

In clearPriorVersion, instead of crashing with FATAL if the CAS fails,
it will just return NULL so the pruning for that record will be skipped.
This change in behavior might prevent a race condition.