Bug #36993 Falcon reports Index SCHEDULE..PRIMARY_KEY in SYSTEM.SCHEDULE damaged
The problem here is that mysqld was killed before database was completely created 
(i.e before all data dictionary was completely written to the disk). Falcon cannot
 handle such sutuations gracefully yet, recovery after such point is not guaranteed
to succeed.

The patch improves the sutation a little bit, disabling  user queiries until database is
fully created and written to the disk. 

Also, this patch introduces a clean Falcon shutdown : waiting for background theads
to complete  their work , followed by flushing the page cache. This will  eliminate the
need for recovery after  a clean shutdown.