Bug#41902  MYSQL_BIN_LOG::reset_logs() doesn't call my_error() in face of an error
The assert happened because thd->stmt_da->status() reacted on the
fact of a gained error in reset_logs() which was not prepared and therefore
sent but still the fact of the error was propagated upwards to callers.

Fixed with deploying a part the former purge_error_message() in the reported error
branch of reset_logs(). A user level error gets prepared for reporting via my_message().
      
Side effects: Bug#44179 appears be fixed with the current patch, at least in 6.0.