Bug #41840 Unbounded tablespace growth during recovery Sometimes, after recovery the tablespaces get huge, orders of magnitude larger than original tablespace.
The problem was that page overflow pages were not deleted 
during recovery. If a record was modified several times during 
recovery, the overflow pages that were created during recovery
also leaked.

The fix is to remove overflow pages , when it is possible.
- remove all SRLOverflowPages found in course of recovery
- just like during normal runtime,when record gets modified in 
recovery,  remove overflow pages if the removal is safe.
Safe means,  that page referenced is a valid dataoverflow page.
Validity is guaranteed, if page is either not touched by recovery, 
or the page was allocated during recovery and not yet freed.